VariLC: Eliminate a warning on 64-bit.
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@11786 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/VariLC/VariLC.cpp
@@ -624,7 +624,7 @@ int VariLC::OnSendToVariLC(MM::PropertyBase* pProp, MM::ActionType eAct)
       pProp->Get(sendToVariLC_);
       // write retardance out to device....
 	  
-	  int len = strlen(sendToVariLC_.c_str());
+	  size_t len = strlen(sendToVariLC_.c_str());
 	  char state[5]; 
 
 	  if (len > 5) {
